+++ src/html/config.mli
@@ -1,16 +1,6 @@
 (* HTML renderer configuration *)
 
-type t = {
-  semantic_uris : bool;
-  indent : bool;
-  flat : bool;
-  open_details : bool;
-  (* only for HTML page rendering*)
-  theme_uri : Types.uri option;
-  support_uri : Types.uri option;
-  (* output HTML fragments and metadata in JSON format *)
-  as_json : bool;
-}
+type t
 
 val v :
   ?theme_uri:Types.uri ->
@@ -23,6 +13,16 @@ val v :
   unit ->
   t
 
-val get_theme_uri : t -> Types.uri
+val theme_uri : t -> Types.uri
 
-val get_support_uri : t -> Types.uri
+val support_uri : t -> Types.uri
+
+val semantic_uris : t -> bool
+
+val indent : t -> bool
+
+val flat : t -> bool
+
+val open_details : t -> bool
+
+val as_json : t -> bool